split project submodule
