+++ skore/src/skore/project/__init__.py
@@ -1,0 +1,11 @@
+"""Alias top level function and class of the project submodule."""
+
+from .create import create
+from .load import load
+from .project import Project
+
+__all__ = [
+    "create",
+    "load",
+    "Project",
+]
